errors: split into multiple go files
